Attempt to fix retry loop when many unauth'd clients rapidly join
8 unauth'd clients all joining in the 1 minute period before the first got retried resulted in a 15 minute retry loop for all new joining (unauth'd) clients.

This likely broke because each players retries were out of sync with each other, so when each client rejoined, there was already someone else blocking their auth, and so on.

Now once a certain amount of clients (configurable) are unauthorized in total, the plugin will switch to a simultaneous retry for everyone instead.